+++ backend/handlers/config.go
@@ -114,7 +114,7 @@ func (h *ConfigHandler) HandleConfig(w http.ResponseWriter, r *http.Request) {
 		"liveCollab":   h.liveCollab,
 		"require2FA":   h.require2FA,
 		"userProvider": h.userProvider,
-		"version":      "4.1.2",
+		"version":      "4.1.3",
 		"commit":       Commit,
 		"buildTime":    BuildTime,
 	}